fix: typography blocks
